% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nivel_asistencia.R
\name{nivel_asistencia}
\alias{nivel_asistencia}
\title{Clasifica el nivel de asistencia}
\usage{
nivel_asistencia(asistencia)
}
\arguments{
\item{asistencia}{Un numero o vector numerico que representa la asistencia
del estudiante, expresada como una proporcion entre 0 y 1.}
}
\value{
Un vector de caracteres con el nivel de asistencia:
"Alta", "Media" o "Baja".
}
\description{
Esta funcion clasifica el nivel de asistencia de un estudiante a partir
de un valor numerico entre 0 y 1. La asistencia no puede contener valores
NA y debe expresarse como proporcion.
}
\examples{
nivel_asistencia(0.9)
nivel_asistencia(0.65)
nivel_asistencia(c(1, 0.75, 0.5))
}
